Decompose core_utils.ts
