+++ src/core/base_types.ts
@@ -1,0 +1,14 @@
+// Copyright (c) 2023 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at https://mozilla.org/MPL/2.0/.
+
+export enum SeedType {
+  PRODUCTION,
+  STAGING,
+  UPSTREAM,
+}
+
+export class ProcessingOptions {
+  minMajorVersion: number;
+}
